Update main page visuals, adjusting buttons
Fleshing out main page template. "Try Demo" and "Upload Spoiler" buttons correctly get the spoiler JSON for us, now we're just trying to figure out how to run the parser on values we get back from them.

Only pushing this so I can continue to work on it while I'm out of town this weekend.

Main page visuals are not finalized, but we're going to close the open issue for now. It's good enough for until we get the graph showing - we can adjust it further once we get there.

closes #16