Demo button runs parser; preparing to switch to css grids
Got our "Try Demo!" button to run the entrance data through the parser after requesting it from the github repo. Next steps:
1. Change the output of the Parser to be more aligned with what D3 needs to generate a network graph
2. Learn how to use D3 to generate a network graph from the output of the parser

Also, flexbox is seemingly the wrong tool to organize the high-level layout of our page - we're going to switch to CSS grids for the big-picture elements, then use flexbox to align elements inside of those.

Pushing these changes so that if/when we break everything in the page layout, we can at least come back to this point.