Update readme, fixes to parser, initial mapper
Parser is outputting json data in the correct format, now we just need to fix the logic of the parser a bit to make the data correct.

The mapper we just copy/pasted from a D3 example on observable, so I don't expect it to be exactly what we want. Really I just want to (once I fix the parser output) be able to look at the output and visually check that the edge assignments match the spoiler correctly - I'll deal with the appearance/properties of the nodes/edges once I get something rendering.

Also updated the gitignore and added some shields.io badges to the readme - probably too many, but whatever I like them lol

closes #23