First output from mapper, various other changes
Mapper generates data and renders it on the webpage. The map is incorrect because the output of the Parser is incomplete, but it runs.

Added additional unit tests for Parser.AdjustNodeNames().

Added top nav bar; added Instructions and About pages.

Initial work on simplifying the appearance of the main page.

closes #13