First correct output from mapper
Style is not finished - all the nodes are the same shape and size, but the output appears correct at first glance.

Map physics still need to be fixed when zoomed in. Map colors should be adjusted further. Spacing should be adjusted to make map more readable.

Also I didn't mean to make these changes on the main branch. Oops.

closes #20, closes #18